oacz-dubbing-job/requirements.txt
google-cloud-storage
google-genai